softirq: Sanitize local_bh_[en|dis]able for RT
There is no point in doing that check for disabling/enabling
preemption in __local_bh_disable_ip/__local_bh_enable_ip on RT because
PREEMPT_CHECK_OFFSET is 0 on RT so the check is always false.

The raw_spin_lock variant is never used on RT because nothing ever uses
raw_spin_*_bh() on RT.

Simplify the rest of the interfaces as well.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>