printk: Do not disable preemption calling console_unlock() in PREEMPT_RT
A stable backport placed open coded preempt_disable() around console_unlock()
to prevent delays in output of printk() due to a task holding the console
lock being preempted. But most consoles are not PREEMPT_RT safe, thus
it needs to be preemptable when PREEMPT_RT is enabled.

Replace the open coded preempt_disable() with migrate_disable() which
is a preempt_disable() on non PREEMPT_RT, but just disables migration
on PREEMPT_RT, even though it doesn't really depend on that.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>