genirq: Do not call cancel_work on old_notify->work on PREEMPT_RT
When PREEMPT_RT is configured, old_notify->work does not exist, and
instead swork is used. Unfortunately, there is no cancel_work() for
swork. This is a known bug, and 4.4 will just need to deal with it
as the real solution to this is to backport all the work queue code
from RT mainline.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>