i2c: rcar: Allow interrupt handler to be threaded.
This is a revert of commit
   24c6d4bc56388 ("i2c: rcar: make sure irq is not threaded on Gen2 and earlier")

The IRQ-handler is not threaded unless requested. On PREEMPT_RT the
handler must be threaded because the handler acquires spinlock_t locks
which are sleeping locks on PREEMPT_RT. This is either via the DMA API
or the possible wake_up() invocation.

Remove IRQF_NO_THREAD from flags passed to request_irq().

Fixes: 24c6d4bc56388 ("i2c: rcar: make sure irq is not threaded on Gen2 and earlier")
Cc: Wolfram Sang <wsa+renesas@sang-engineering.com>
Cc: linux-renesas-soc@vger.kernel.org
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>